Application/requirements.txt
fastapi
uvicorn
diffusers
transformers
accelerate
safetensors
torch
torchvision 
torchaudio
python-dotenv
python-multipart
huggingface_hub